feat: add a document list view
* separate layout into components (header / body / footer)
* add a document list component
* mock doc list based on static data
