+++ src/index.tsx
@@ -2,7 +2,6 @@ import * as React from 'react'
 import ReactDOM from 'react-dom'
 import { App } from './App'
 import { QueryClient, QueryClientProvider } from 'react-query'
-import { Redirect, Router } from '@reach/router'
 import theme from './theme'
 import { ChakraProvider, ColorModeScript } from '@chakra-ui/react'
 
@@ -13,14 +12,7 @@ ReactDOM.render(
     <QueryClientProvider client={queryClient}>
       <ChakraProvider theme={theme}>
         <ColorModeScript initialColorMode={theme.config.initialColorMode} />
-        <Router>
-          <Redirect
-            from="/*"
-            to="/document/k3y52l7qbv1frxjdr9qpn9ldvbxb0jg4eig7wtjkdu6gk84vyazw9j4txf4o6d2io"
-            noThrow
-          />
-          <App path="/document/:docId" />
-        </Router>
+        <App />
       </ChakraProvider>
     </QueryClientProvider>
   </React.StrictMode>,
